Remove Zero bound for transpose and impl SparsityPattern::transpose
